% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.p.sig.mod.R, R/pcps.sig.mod.R,
%   R/mutate.names.matrix.p.null.R
\encoding{UTF-8}
\name{matrix.p.sig.mod}
\alias{matrix.p.sig.mod}
\alias{pcps.sig.mod}
\alias{mutate.names.matrix.p.null}
\title{Internal functions}
\usage{
matrix.p.sig.mod(comm, phylodist, FUN, runs = 999, parallel = NULL,
  ...)

pcps.sig.mod(comm, phylodist, method = "bray", squareroot = TRUE, FUN,
  choices, runs = 999, parallel = NULL, ...)

mutate.names.matrix.p.null(x, replacement, newname)
}
\arguments{
\item{comm}{Community data, with species as columns and sampling units as rows.}

\item{phylodist}{Matrix containing phylogenetic distances between species.}

\item{FUN}{An object of class function to perform the analysis.}

\item{runs}{Number of permutations for assessing significance.}

\item{parallel}{Number of parallel processes or a predefined socket cluster done with parallel package.}

\item{...}{Other arguments passed to FUN function.}

\item{method}{Dissimilarity index.}

\item{squareroot}{Logical argument (TRUE or FALSE) to specify if use square root of dissimilarity index.}

\item{choices}{Numeric vector to choose the axis used in analysis.}

\item{x}{An object returned by \code{\link{matrix.p.null}}.}

\item{replacement}{A replacement name to matched in object returned by \code{\link{matrix.p.null}}.}

\item{newname}{New name to be replaced in object returned by \code{\link{matrix.p.null}}.}
}
\description{
Internal functions with small modifications of \code{\link{matrix.p.sig}} and \code{\link{pcps.sig}}.
}
