\encoding{UTF-8}
\name{segv}

\alias{segv}

\docType{data}

\title{segv - Simple Example for GenVectors}

\description{Hypothetical data for running examples.}

\usage{data(segv)}

\format{
A list with: 
\describe{\item{segv.fas}{An object of class DNAbin with 18 DNA sequences.} \item{segv.pi}{A matrix describing the incidence of each individual (columns) in a given locality (rows).}  \item{segv.envir}{A matrix with one environmental variable describing each locality.} 
}
}

\examples{
data(segv)
segv
}

\keyword{datasets}