\encoding{UTF-8}
\name{sehv}

\alias{sehv}

\docType{data}

\title{sehv - Simple Example for HaploVectors}

\description{Hypothetical data for running examples.}

\usage{data(sehv)}

\format{
A list with: 
\describe{\item{sehv.fas}{An object of class DNAbin with 18 DNA sequences.} \item{sehv.pi}{A matrix describing the incidence of each individual (columns) in a given locality (rows).}  \item{sehv.envir}{A matrix with one environmental variable describing each locality.} 
}
}

\examples{
data(sehv)
sehv
}

\keyword{datasets}