% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPDist.R
\encoding{UTF-8}
\name{SNPDist}
\alias{SNPDist}
\title{SNPDist}
\usage{
SNPDist(x)
}
\arguments{
\item{x}{A set of individual genotypes (an object of class \code{\link{genind}}) as used by the function \code{\link{propShared}}.}
}
\value{
A list with: \item{call}{Arguments used.} 
\item{SNP.distances}{A matrix of distances between alleles.}
}
\description{
Function to computes distances between alleles. This function use the function \code{\link{propShared}} of package adegenet.
}
\examples{
data(segv)
}
\seealso{
\code{\link{SNPVectors}}
}
