% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaploNetDist.R
\encoding{UTF-8}
\name{HaploNetDist}
\alias{HaploNetDist}
\title{HaploNetDist}
\usage{
HaploNetDist(x, dist.model = "N", ...)
}
\arguments{
\item{x}{A set of DNA sequences (as an object of class "DNAbin" or "haplotype") as used by the function \code{\link{haplotype}}.}

\item{dist.model}{A character string used by the function \code{\link{dist.dna}} to specify the evolutionary model to be used to compute pairwise distances from DNA sequences (default dist.model = "N").}

\item{...}{Additional arguments to the function \code{\link{dist.dna}}.}
}
\value{
A list with: \item{call}{Arguments used.} 
\item{haplotypes}{A list with haplotypes indices that identify each observation sharing the same haplotype.} 
\item{individual.per.haplotype}{A matrix with individuals per haplotype.} 
\item{haplotype.distances}{A matrix with pairwise distances between haplotypes.}
}
\description{
Function to extract haplotypes and compute pairwise distances between haplotypes.
}
\examples{
data(segv)
HaploNetDist(segv$segv.fas)
}
\seealso{
\code{\link{HaploVectors}}
}
