% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaploVectors.R
\encoding{UTF-8}
\name{HaploVectors}
\alias{HaploVectors}
\title{HaploVectors}
\usage{
HaploVectors(x, pop, dist.model = "N", log.frequencies = TRUE,
  squareroot.network = TRUE, method = "euclidean", squareroot.dis = TRUE,
  choices = c(1, 2), analysis = "none", envir, formula, AsFactors = NULL,
  runs = 999, ...)
}
\arguments{
\item{x}{A set of DNA sequences (as an object of class "DNAbin" or "haplotype") as used by the function \code{\link{haplotype}}.}

\item{pop}{A matrix describing the incidence of each individual (columns) in a given locality (rows).}

\item{dist.model}{A character string used by the function \code{\link{dist.dna}} to specify the evolutionary model to be used to computes pairwise distances from DNA sequences (default dist.model = "N").}

\item{log.frequencies}{Logical argument (TRUE or FALSE) to specify if transformation of natural logarithms plus one in haplotype per locality data must be applied  (Default log.frequencies = TRUE).}

\item{squareroot.network}{Logical argument (TRUE or FALSE) to specify if apply square root on haplotype network (Default squareroot.network = TRUE).}

\item{method}{Dissimilarity index to apply in matrix P, which describes localities by their haplotypic composition, as accepted by vegdist function in vegan package (Default method = "euclidean").}

\item{squareroot.dis}{Logical argument (TRUE or FALSE) to specify if use square root of dissimilarity index in matrix P (Default squareroot.dis = TRUE).}

\item{choices}{Axes for re-scaling. Choices must have length equal to two (Default choices = c(1, 2)).}

\item{analysis}{Type of analysis, partial match to "none", "glm", "adonis" (Default analysis = "none").}

\item{envir}{A matrix with environmental variables for each population, with variables as columns and localities as rows.}

\item{formula}{An object of class formula with quotation marks, used in GLM analysis. See Details.}

\item{AsFactors}{Encode an environmental variable as factor used in GLM analysis. The sequence is the same that in the environmental data matrix. See Details.}

\item{runs}{Number of permutations for assessing probability of type I error.}

\item{...}{Aditional arguments to function \code{\link{matrix.p.sig}} and \code{\link{pcps.sig}}.}
}
\value{
A list with: \item{call}{Arguments used.} 
\item{haplotypes}{A list with haplotypes index that identify each observation that share the same haplotype.} 
\item{haplotype.distances}{A matrix with pairwise distances between haplotypes.} 
\item{haplotype.per.locality}{A matrix with frequency of each haplotype per locality (\bold{\eqn{W}}).} 
\item{haplotype.network}{A matrix with haplotype connections described in the network (\bold{\eqn{D_n}}).} 
\item{vectors}{Haplotypic eigenvectors (haplovectors).} 
\item{values}{Eigenvalues, relative eigenvalues and cumulative relative eigenvalues.} 
\item{correlations}{Correlations between haplotypic eigenvectors and haplotypes.} 
\item{P}{Matrix of haplotypic composition (\bold{\eqn{P}}).} 
\item{scores.haplotypes}{Scores for biplots.} 
\item{envir.class}{The class of each variable in environmental data in glm.}
\item{formula}{The formula used in glm.} 
\item{model}{The model, an object of class glm or adonis.}
\item{statistic.obs}{Observed F value.}
\item{p.site.shuffle}{The p value for the site shuffle null model.}
\item{p.taxa.shuffle}{The p value for the taxa shuffle null model.}
}
\description{
Function to extract haplotypic eigenvectors and perform null model-based tests.
}
\details{
The item formula is an expression of the form haplovector.1 ~ envir. The response term must be the 
haplovector name, for example, haplovector.1, haplovector.2, haplovector.12, etc.

The item AsFactors changes an environmental variable for the class \code{\link{factor}}. The 
sequence is the same that in the environmental data matrix, not the order in the formula.
Use \code{\link{c}} to combine more than one variable.
}
\examples{
data(sehv)
HaploVectors(sehv$sehv.fas, sehv$sehv.pi, envir = sehv$sehv.envir)
HaploVectors(sehv$sehv.fas, sehv$sehv.pi, envir = sehv$sehv.envir, 
             analysis = "glm", runs = 99, formula = haplovector.1~R)
}
\seealso{
\code{\link{HaploNetDist}}
}
